requirements.txt
fastapi
uvicorn
motor
bcrypt==4.0.1
passlib==1.7.4
pydantic-settings
email-validator
google-auth
pymongo